Feature: Added a way to load .mine maps. Feature: Added confirmation for potentially destructive commands: /importranks, /importbans, /massrank, /wload overwrites, /gen overwrites, /wsave overwrites, /zremove. Feature: Added a way to specify default build permission for new worlds. Fix: Replaced File.Delete/File.Move calls with a more reliable File.Replace call. This should slightly improve reliability of map/playerdb.txt/ipbans.txt/worlds.xml saving. Fix: Fixed a little ConfigTool bug that did not properly reset PatrolledRank after a rank deletion.

+++ trunk/fCraft/Network/IPBanList.cs
@@ -55,8 +55,11 @@ namespace fCraft {
                 }
             }
             try {
-                File.Delete( BanFile );
-                File.Move( tempFile, BanFile );
+                if( File.Exists( BanFile ) ) {
+                    File.Replace( tempFile, BanFile, null, true );
+                } else {
+                    File.Move( tempFile, BanFile );
+                }
             } catch( Exception ex ) {
                 Logger.Log( "IPBanList.Save: An error occured while trying to save ban list file: " + ex, LogType.Error );
             }
